Cleaning codebase
