--- apps/tsforjs/src/components/original/MainContent/CourseTitles.js
@@ -1,54 +1,0 @@
-export const Titles = [
-  {
-    number: 1,
-    title: 'Why TypeScript',
-  },
-  {
-    number: 2,
-    title: 'Configuring Your Environment',
-  },
-  {
-    number: 3,
-    title: 'Intro to Types',
-    content: [
-      'Type Inference',
-      'Basic Types',
-      'Object Types',
-      'Enums',
-      'Function Types',
-    ],
-  },
-  {
-    number: 4,
-    title: 'Type Aliases, Unions, and Intersections',
-    content: ['Type Aliases', 'Interfaces', 'Unions', 'Intersections'],
-  },
-]
-
-export const TitlesTwo = [
-  {
-    number: 5,
-    title: 'Advanced Types',
-    content: [
-      'Conditional Types',
-      'Template Literal Types',
-      'Type Operators',
-      'Generics',
-      'Call Signatures',
-      'Type-Driven Development',
-    ],
-  },
-  {
-    number: 6,
-    title: 'Error Handling',
-  },
-  {
-    number: 7,
-    title: 'Real-World Application Development',
-    content: [
-      'Configuration',
-      'Building from Scratch',
-      'Converting to TypeScript from JavaScript',
-    ],
-  },
-]
